Remove existing validator subcommands.
Add renamed, hidden commands to new internal-tools subcommand.  Add
warnings when these commands are invoked and update descriptions to warn
against production use.
